requirements.txt
Django>=3.2
djangorestframework
pandas
openpyxl
djangorestframework.authtoken
rest_framework_simplejwt